fix(lsm): widen write seq to u64, version SSTables, close two crash/corruption holes
Five related changes to the LSM storage layer. The first two are a fix and the
guard it required; the last two are pre-existing bugs found while auditing it.

1. Write sequence is u64 end-to-end (was silently truncated to u32).

`KVStore::alloc_seq` truncated the global u64 counter to u32, and the LSM
compared with wrapping RFC-1982 serial arithmetic — valid only within 2^31.
That reasoning holds for the memtable (it flushes long before 2B writes) but
not for the SSTable levels: an L1 entry keeps its seq for the lifetime of the
database, so the gap to a later write of the same key is unbounded. Past ~2.1B
lifetime writes (~2.5 days at 10k writes/s) updating a cold key made the newer
write compare as older — `get` returned the stale L1 value and the L0->L1 merge
then baked it into L1 permanently, silently.

Widened Node.sequence, KeyValueRecord.seq, SStableEntry.seq, SsLookup and
max_lower_seq; seq_newer/seq_newer_or_eq/seq_is_newer_or_equal are now plain
comparisons. The WAL and value log already stored the seq as u64, so the LSM
was the only narrowing point. Free in memory — u128 alignment already padded
both structs (Node 48B, SStableEntry 64B, unchanged).

2. SSTable files carry a magic + format version.

The widening changed the on-disk entry layout, and the per-entry CRC covers only
the body — it cannot detect a layout change, because the bytes are identical and
only their interpretation differs. An old file would pass CRC and then be handed
to rkyv::access_unchecked under the new layout. Every non-empty SSTable (L0 and
L1) now starts with [magic "MSST" | format_version u32 | reserved], validated at
open. A database written before this must be recreated, but now fails loudly.

Headers are validated once per file, not per read: a file's format is fixed when
it is written, L0 files are write-once, and both registry entry points establish
it. Re-validating cost +147ns per L0 file access (+28%) on every point get, since
search_level0_files opens every L0 file in the bucket.

3. L1 publish is a single atomic rename (pre-existing data loss).

merge_level0_to_level1 moved the live file aside (rename level1 -> .old) before
renaming the new one in. A crash between the two left the L1 path absent;
LSMTree::open then created a 0-byte file, file_len == 0 short-circuited the
metadata load into "empty bucket", and cleanup_old_files_on_startup deleted the
.old holding the only surviving copy. Every key that lived only in L1 was lost,
silently, with a complete .old and .tmp both sitting on disk unread.

rename(2) already replaces the destination atomically, so the intermediate step
was never needed. Also fsync the L1 directory so the swap itself is durable.
Reproduced identically at HEAD before the change — not caused by items 1-2.

4. Frame lengths are bounded before being allocated from (pre-existing).

The 4-byte length prefix sits outside the entry CRC, so a flipped bit there is
undetectable until after the payload is read — and readers allocated from it
first, letting a corrupt length request up to 4 GiB from a file of a few dozen
bytes. check_frame_len now rejects a frame longer than its file at all 13
sequential read sites (valid_scan_start already had a stricter inline form).

Tests: +2 regression tests, both verified to fail without their fix.
test_crash_between_l1_publish_steps_keeps_l1 watches the L1 path from a second
thread across 200 compactions — a final-state assertion is vacuous, since a
completed merge removes its own .old (the first version of that test passed
against the bug). test_corrupt_frame_length_is_rejected_before_allocating
asserts the guard's own error, because an unguarded reader also errors, just as
Io(UnexpectedEof) after committing the allocation.

1195 tests pass; clippy -D warnings clean across default / doc-store /
semantic-search / --all-features and the full workspace.

+++ minnal_db/src/db/kv_store.rs
@@ -31,8 +31,8 @@ use std::time::Duration;
 /// A resolved key/value pair: both the key and the value bytes read from the value log.
 pub(crate) type KeyValue = (Vec<u8>, Vec<u8>);
 
-/// A key paired with its encoded value-log pointer and the LSM write `seq` (low u32).
-pub(crate) type KeyPointer = (Vec<u8>, u128, u32);
+/// A key paired with its encoded value-log pointer and the full-width LSM write `seq`.
+pub(crate) type KeyPointer = (Vec<u8>, u128, u64);
 
 /// One page of a cursor scan: the resolved pairs plus the next cursor
 /// (`None` when the scan is exhausted).
@@ -295,10 +295,15 @@ impl KVStore {
         *self.seq_counter.write() = counter;
     }
 
-    /// Allocate the next write sequence (truncated to the skip list's `u32`
-    /// sequence width; serial-number comparison tolerates the truncation).
-    pub(crate) fn alloc_seq(&self) -> u32 {
-        self.seq_counter.read().fetch_add(1, Ordering::Relaxed) as u32
+    /// Allocate the next write sequence at full `u64` width.
+    ///
+    /// Must not be narrowed: the LSM compares a fresh write's sequence against
+    /// entries that can be arbitrarily old (an L1 entry keeps its sequence for
+    /// the lifetime of the database), so the width has to span the whole
+    /// sequence space, not just a recent window. See [`crate::store::lsm::lsm_tree`]'s
+    /// `seq_newer`.
+    pub(crate) fn alloc_seq(&self) -> u64 {
+        self.seq_counter.read().fetch_add(1, Ordering::Relaxed)
     }
 
     // ── Row-ID resolution ──────────────────────────────────────────────
@@ -437,7 +442,7 @@ impl KVStore {
     /// uses [`put_to_storage_seq`](Self::put_to_storage_seq) so the conflict-
     /// resolution sequence equals the WAL append order (live == recovery).
     pub fn put_to_storage(&self, key: &[u8], value: &[u8]) -> Result<()> {
-        self.put_to_storage_inner(key, value, self.alloc_seq() as u64)
+        self.put_to_storage_inner(key, value, self.alloc_seq())
     }
 
     /// Put a key-value pair carrying an explicit global write sequence, so
@@ -464,7 +469,7 @@ impl KVStore {
         let bucket = self.value_log.bucket_for_key(key);
         let _bucket_guard = self.value_log.lock_bucket_for_write(bucket)?;
 
-        let (existing_u128, wins) = self.lsm.current_pointer_and_wins(key, seq as u32)?;
+        let (existing_u128, wins) = self.lsm.current_pointer_and_wins(key, seq)?;
         let displaced = existing_u128.and_then(decode_sharded_pointer);
         let mut old_value: Option<Vec<u8>> = None;
         if let Some(existing_ptr) = displaced {
@@ -482,7 +487,7 @@ impl KVStore {
             seq,
         };
         let pointer = self.value_log.append_to_locked_bucket(bucket, key, value, record_meta, false)?;
-        self.lsm.insert_with_seq(key, pointer.to_u128(), seq as u32)?;
+        self.lsm.insert_with_seq(key, pointer.to_u128(), seq)?;
 
         // Account the record that is now garbage — accounting only, no I/O and no
         // in-place edit (the old record's segment may be sealed and immutable). If this
@@ -689,7 +694,7 @@ impl KVStore {
     /// store's counter (used by TTL expiry / standalone tests); the WAL-backed
     /// path uses [`delete_from_storage_seq`](Self::delete_from_storage_seq).
     pub fn delete_from_storage(&self, key: &[u8]) -> Result<()> {
-        self.delete_from_storage_inner(key, self.alloc_seq() as u64)
+        self.delete_from_storage_inner(key, self.alloc_seq())
     }
 
     /// Delete carrying an explicit global write sequence, so a delete racing a
@@ -724,7 +729,7 @@ impl KVStore {
                 old_value = self.value_log.read_value(existing_ptr).ok();
             }
         }
-        self.lsm.delete_with_seq(key, seq as u32)?;
+        self.lsm.delete_with_seq(key, seq)?;
 
         // The deleted key's record is garbage now. Accounting only — the record itself
         // is never touched (its segment may be sealed and immutable), and liveness was
@@ -1182,7 +1187,7 @@ impl KVStore {
         let bucket = self.value_log.bucket_for_key(key);
         let _bucket_guard = self.value_log.lock_bucket_for_write(bucket)?;
 
-        let (existing_u128, wins) = self.lsm.current_pointer_and_wins(key, seq as u32)?;
+        let (existing_u128, wins) = self.lsm.current_pointer_and_wins(key, seq)?;
         let displaced = existing_u128.and_then(decode_sharded_pointer);
         if let Some(existing_ptr) = displaced
             && let Ok(meta) = self.value_log.read_record_meta(existing_ptr)
@@ -1196,7 +1201,7 @@ impl KVStore {
             seq,
         };
         let pointer = self.value_log.append_to_locked_bucket(bucket, key, value, record_meta, false)?;
-        self.lsm.insert_with_seq(key, pointer.to_u128(), seq as u32)?;
+        self.lsm.insert_with_seq(key, pointer.to_u128(), seq)?;
         if wins {
             if let Some(old) = displaced {
                 self.value_log.note_displaced(old, key.len());
@@ -1218,7 +1223,7 @@ impl KVStore {
         let bucket = self.value_log.bucket_for_key(key);
         let _bucket_guard = self.value_log.lock_bucket_for_write(bucket)?;
         let displaced = self.lsm.get(key).ok().flatten().and_then(decode_sharded_pointer);
-        self.lsm.delete_with_seq(key, seq as u32)?;
+        self.lsm.delete_with_seq(key, seq)?;
         if let Some(old) = displaced {
             self.value_log.note_displaced(old, key.len());
         }
